Fix JSON parser constructor
:Release Notes:
Add missing member initialization in JSON parser constructor.

:Detailed Notes:
The constructor of the JSON parser was missing initialization of
the `_valueNull` member. This patch adds `_valueNull` to the list
of initialized members in the constructor, ensuring all members
are correctly initialized from the provided context.

:Testing Performed:
Test cases verified: QWTC-9080, QWTC-10388, QWTC-8560

:QA Notes:
N/A

:Issues Addressed:
[WRR-18303] Patch merge for ls2-helper in repo

Change-Id: Ie480ba5f94b74b6ef6564d50073be232d7ebdd4a